+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## 0.38.15
 
+* `Core`:
+    * Fixes in `MessageContent#serializationModule`
+
 ## 0.38.14
 
 __This update contains including of [Telegram Bot API 6.0](https://core.telegram.org/bots/api-changelog#april-16-2022)__